img-checker/requirements.txt
streamlit>=1.36
opencv-python>=4.9
Pillow>=10.3
numpy>=1.26
scipy>=1.12
